auto_ui/requirements.txt
allure-pytest==2.8.19
allure-python-commons==2.8.19
PyMySQL==0.9.3
pytest==6.1.2
pytest-assume==2.3.3
pytest-forked==1.3.0
pytest-ordering==0.6
pytest-repeat==0.9.1
pytest-rerunfailures==9.1.1
pytz==2020.1
requests==2.24.0
selenium==3.141.0